requirements.txt
Flask==2.3.2

Flask-SQLAlchemy==3.0.3

mysql-connector-python==8.0.33

Werkzeug==2.3.7
python-dotenv==1.0.0 
Flask-WTF==1.1.1